Remove CallbackInfo, closes #67
