+++ stubs/src/PropertyCallbackInfo.php
@@ -19,8 +19,56 @@ namespace V8;
  * The information passed to a property callback about the context
  * of the property access.
  */
-class PropertyCallbackInfo extends CallbackInfo
+class PropertyCallbackInfo
 {
+    /**
+     * @return Isolate
+     */
+    public function getIsolate(): Isolate
+    {
+    }
+
+    /**
+     * @return Context
+     */
+    public function getContext(): Context
+    {
+    }
+
+    /**
+     * Returns the receiver. This corresponds to the "this" value.
+     *
+     * @return ObjectValue
+     */
+    public function this(): ObjectValue
+    {
+    }
+
+    /**
+     * If the callback was created without a Signature, this is the same
+     * value as This(). If there is a signature, and the signature didn't match
+     * This() but one of its hidden prototypes, this will be the respective
+     * hidden prototype.
+     *
+     * Note that this is not the prototype of This() on which the accessor
+     * referencing this callback was found (which in V8 internally is often
+     * referred to as holder [sic]).
+     *
+     * @return ObjectValue
+     */
+    public function holder(): ObjectValue
+    {
+    }
+
+    /**
+     * The ReturnValue for the call
+     *
+     * @return ReturnValue
+     */
+    public function getReturnValue(): ReturnValue
+    {
+    }
+
     /**
      * Returns true if the intercepted function should throw if an error occurs.
      * Usually, true corresponds to 'use strict'.
