+++ tests/001-verify_extension_entities.phpt
@@ -884,28 +884,34 @@ class V8\ReturnValue
     public function getContext(): V8\Context
     public function inContext(): bool
 
-class V8\CallbackInfo
+class V8\PropertyCallbackInfo
     private $isolate
     private $context
     private $this
     private $holder
     private $return_value
+    private $should_throw_on_error
     public function this(): V8\ObjectValue
     public function holder(): V8\ObjectValue
     public function getIsolate(): V8\Isolate
     public function getContext(): V8\Context
     public function getReturnValue(): V8\ReturnValue
-
-class V8\PropertyCallbackInfo
-    extends V8\CallbackInfo
-    private $should_throw_on_error
     public function shouldThrowOnError(): bool
 
 class V8\FunctionCallbackInfo
-    extends V8\CallbackInfo
+    private $isolate
+    private $context
+    private $this
+    private $holder
+    private $return_value
     private $arguments
     private $new_target
     private $is_constructor_call
+    public function this(): V8\ObjectValue
+    public function holder(): V8\ObjectValue
+    public function getIsolate(): V8\Isolate
+    public function getContext(): V8\Context
+    public function getReturnValue(): V8\ReturnValue
     public function length(): int
     public function arguments(): array
     public function newTarget(): V8\Value
